Add functions
Very big jump from having one smaller function to the entire project done, but here it is. Essentially, there are three functions: one for the player, one for the computer, and one for the actual game itself. The player function gives the user a prompt to type rock, paper, or scissors and parse it into lowercase. The computer function generates a number between 1 and 3, then chooses either Rock, Paper, or Scissors accordingly. The game function compares the results of the two previous functions and has six outcomes.